Slow down post removal polling interval from 1s to 2s
This may be premature as haven't tested on a real device yet, but
it feels slow in the emulator.

Thought: Maybe it feels slow though not because of performance, but
because some scroll-to-load actions are loading only more posts
from already-encountered authors. This would cause no change in the
UI then and it looks like the app isn't responding. Possible solution:
show a counter hovering in the bottom or corner of the UI showing
total number of posts that have been hidden so far.